requirements.txt
pandas==2.0.2
psycopg2_binary==2.9.6
python-dotenv==1.0.0
streamlit==1.24.0
